move ServiceLegend.jsx under RasterStyle
